PositiveIntegrators.jl API

PositiveIntegrators.ConservativePDSProblem — Type

ConservativePDSProblem(P, u0, tspan, p = NullParameters();
                        p_prototype = nothing, 
                        analytic=nothing)

A structure describing a conservative system of ordinary differential equation in form of a production-destruction system (PDS). P denotes the production matrix. u0 is the vector of initial conditions and tspan the time span (t_initial, t_final) of the problem. The optional argument p can be used to pass additional parameters to the function P.

The function P can be given either in the out-of-place form with signature production_terms = P(u, p, t) or the in-place form P(production_terms, u, p, t).

Keyword arguments:

p_prototype: If P is given in in-place form, p_prototype is used to store evaluations of P. If p_prototype is not specified explicitly and P is in-place, then p_prototype will be internally set to zeros(eltype(u0), (length(u0), length(u0))).

analytic: The analytic solution of a PDS must be given in the form f(u0,p,t). Specifying the analytic solution can be useful for plotting and convergence tests.

References

Hans Burchard, Eric Deleersnijder, and Andreas Meister. "A high-order conservative Patankar-type discretisation for stiff systems of production-destruction equations." Applied Numerical Mathematics 47.1 (2003): 1-30. DOI: 10.1016/S0168-9274(03)00101-6

PositiveIntegrators.MPE — Type

MPE()

The first-order modified Patankar-Euler algorithm for conservative production-destruction systems. This one-step, one-stage method is first-order accurate, unconditionally positivity-preserving, and linearly implicit.

The modified Patankar-Euler method requires the special structure of a PDSProblem or a ConservativePDSProblem.

References

Hans Burchard, Eric Deleersnijder, and Andreas Meister. "A high-order conservative Patankar-type discretisation for stiff systems of production-destruction equations." Applied Numerical Mathematics 47.1 (2003): 1-30. DOI: 10.1016/S0168-9274(03)00101-6

PositiveIntegrators.MPRK22 — Type

MPRK22(α)

The second-order modified Patankar-Runge-Kutta algorithm for conservative production-destruction systems. This one-step, two-stage method is second-order accurate, unconditionally positivity-preserving, and linearly implicit. The parameter α is described by Kopecz and Meister (2018) and studied by Izgin, Kopecz and Meister (2022) as well as Torlo, Öffner and Ranocha (2022).

This modified Patankar-Runge-Kutta method requires the special structure of a PDSProblem or a ConservativePDSProblem.

References

Hans Burchard, Eric Deleersnijder, and Andreas Meister. "A high-order conservative Patankar-type discretisation for stiff systems of production-destruction equations." Applied Numerical Mathematics 47.1 (2003): 1-30. DOI: 10.1016/S0168-9274(03)00101-6

Stefan Kopecz and Andreas Meister. "On order conditions for modified Patankar-Runge-Kutta schemes." Applied Numerical Mathematics 123 (2018): 159-179. DOI: 10.1016/j.apnum.2017.09.004

Thomas Izgin, Stefan Kopecz, and Andreas Meister. "On Lyapunov stability of positive and conservative time integrators and application to second order modified Patankar-Runge-Kutta schemes." ESAIM: Mathematical Modelling and Numerical Analysis 56.3 (2022): 1053-1080. DOI: 10.1051/m2an/2022031

Davide Torlo, Philipp Öffner, and Hendrik Ranocha. "Issues with positivity-preserving Patankar-type schemes." Applied Numerical Mathematics 182 (2022): 117-147. DOI: 10.1016/j.apnum.2022.07.014

PositiveIntegrators.PDSProblem — Type

PDSProblem(P, D, u0, tspan, p = NullParameters();
                   p_prototype = nothing,
                   d_prototype = nothing,
                   analytic = nothing)

A structure describing a system of ordinary differential equations in form of a production-destruction system (PDS). P denotes the production matrix. The diagonal of P contains production terms without destruction counterparts. D is the vector of destruction terms without production counterparts. u0 is the vector of initial conditions and tspan the time span (t_initial, t_final) of the problem. The optional argument p can be used to pass additional parameters to the functions P and D.

The functions P and D can be used either in the out-of-place form with signature production_terms = P(u, p, t) or the in-place form P(production_terms, u, p, t).

Keyword arguments:

p_prototype: If P is given in in-place form, p_prototype is used to store evaluations of P. If p_prototype is not specified explicitly and P is in-place, then p_prototype will be internally set to zeros(eltype(u0), (length(u0), length(u0))).

d_prototype: If D is given in in-place form, d_prototype is used to store evaluations of D. If d_prototype is not specified explicitly and D is in-place, then d_prototype will be internally

set to zeros(eltype(u0), (length(u0),)).

analytic: The analytic solution of a PDS must be given in the form f(u0,p,t). Specifying the analytic solution can be useful for plotting and convergence tests.

References

Hans Burchard, Eric Deleersnijder, and Andreas Meister. "A high-order conservative Patankar-type discretisation for stiff systems of production-destruction equations." Applied Numerical Mathematics 47.1 (2003): 1-30. DOI: 10.1016/S0168-9274(03)00101-6